+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `hover` option for tooltips
 - Added `html` option for tooltips
+- Added `getMapObject` function
 - Updated default style to `streets-v12`
 - Reduced dependencies
 - Dropped support for AJAX through jQuery and Zepto